feat: move state functions to new file
